move pipe batcher to own module
